sample.csv
A format to upload employees data can be added easily to database is, create a CSV file and upload it to the bot chat and as a replay to the above file, send /load command. This is an example file that can be used in this bot. All the fields can later be modified using individual commands.